+++ packages/imperative/src/cmd/src/doc/handler/ICommandHandlerResponseValidator.ts
@@ -9,7 +9,7 @@
 *
 */
 
-import { ICommandResponse } from "../../../src/doc/response/response/ICommandResponse";
+import { ICommandResponse } from "../../doc/response/response/ICommandResponse";
 import { CommandResponse } from "../../response/CommandResponse";
 
 export interface ICommandHandlerResponseValidator {